requirements.txt
httpx
pydantic
pytest-asyncio
stackprinter
typing_extensions